various: better UTC times
