+++ rero_mef/api_mef.py
@@ -17,10 +17,9 @@
 
 """API for manipulating MEF records."""
 
-from datetime import datetime
+from datetime import datetime, timezone
 
 import click
-import pytz
 from elasticsearch_dsl import Q
 from flask import current_app
 from invenio_search import current_search
@@ -212,23 +211,9 @@ class EntityMefRecord(ReroMefRecord):
         :param reindex: Reindex record.
         :returns: Modified record.
         """
-        self['deleted'] = pytz.utc.localize(datetime.now()).isoformat()
+        self['deleted'] = datetime.now(timezone.utc).isoformat()
         self.update(data=self, dbcommit=dbcommit, reindex=reindex)
         return self
-
-    @classmethod
-    def create_deleted(cls, record, dbcommit=False, reindex=False):
-        """Create a deleted record for an record.
-
-        :param record: Record to create.
-        :param dbcommit: Commit changes to DB.
-        :param reindex: Reindex record.
-        :returns: Created record.
-        """
-        data = {record.name: {
-            '$ref': build_ref_string(record.pid, record.name)}}
-        data['deleted'] = pytz.utc.localize(datetime.now()).isoformat()
-        return cls.create(data=data, dbcommit=dbcommit, reindex=reindex)
 
     @classmethod
     def flush_indexes(cls):
