+++ rero_mef/concepts/mef/api.py
@@ -17,6 +17,8 @@
 
 """API for manipulating MEF records."""
 
+from datetime import datetime, timezone
+
 from flask import current_app
 from invenio_search import current_search
 from invenio_search.api import RecordsSearch
@@ -95,8 +97,7 @@ class ConceptMefRecord(EntityMefRecord):
         for agent in ['rero']:
             if agent in data and data[agent]:
                 sources.append(agent)
-                metadata = data[agent].get('metadata')
-                if metadata:
+                if metadata := data[agent].get('metadata'):
                     data[agent] = metadata
         data['sources'] = sources
         return data
@@ -111,6 +112,21 @@ class ConceptMefRecord(EntityMefRecord):
         :returns: MEF record, MEF action, VIAF record, VIAF
         """
         return self, Action.Error, None, False
+
+    @classmethod
+    def create_deleted(cls, record, dbcommit=False, reindex=False):
+        """Create a deleted record for an record.
+
+        :param record: Record to create.
+        :param dbcommit: Commit changes to DB.
+        :param reindex: Reindex record.
+        :returns: Created record.
+        """
+        data = {
+            record.name: {'$ref': build_ref_string(record.pid, record.name)},
+            'deleted': datetime.now(timezone.utc).isoformat()
+        }
+        return cls.create(data=data, dbcommit=dbcommit, reindex=reindex)
 
 
 class ConceptMefIndexer(ReroIndexer):
